clir-engine/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
scikit-learn==1.4.1.post1
rank-bm25==0.2.2
sentence-transformers==2.6.1
rapidfuzz==3.6.1
faiss-cpu==1.7.4
transformers==4.37.2
torch==2.2.1
sentencepiece==0.1.99
sacremoses==0.1.1
